+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## unreleased
+
+- **BREAKING**: Add type generic to `Edge` class
+
 ## 0.2.0
 
-- Update README.md
-- Make example more usable
 - Implement node pinning with `controller.setPinned`
 - Implement node replacement with `controller.replaceNode`
 - Implement clearing the graph with `controller.clear`
+- Update README.md
+- Make example more usable
 
 ## 0.1.1